Fix gem line for 1.0.0 tag
The example line for using the 1.0.0 gem version didn’t work. It resulted in a couple errors.
First, since it was using `github` it appended `git://github.com` to the front and `.git` the end. This resulted in a bad url.
Second, the tag `1.0.0` is wrong. The correct tag is `v1.0.0`.

To resolve these issues:
I changed the line to use the more standard `git` syntax.
I also fixed the tag name by changing it from `1.0.0` to `v1.0.0`.